delete unused components. add Square component and transfer its code
