--- src/BassDrumRow/BassDrumRow.css
@@ -1,20 +1,0 @@
-.bass-track {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-.bass-note {
-    border: 1px solid black;
-    min-height: 50px;
-    min-width: 50px;
-    cursor: pointer;
-  }
-  
-  .bass-red {
-    background-color: lightcoral;
-  }
-  
-  .bass-green {
-    background-color: lightseagreen;
-  }
