--- src/ClosedHatRow/ClosedHatRow.css
@@ -1,45 +1,0 @@
-.closed-hat-track {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.closed-hat-note {
-  border: 1px solid black;
-  min-height: 50px;
-  min-width: 50px;
-  max-height: 50px;
-  max-width: 50px;
-  cursor: pointer;
-}
-
-.closed-hat-red {
-  background-color: lightcoral;
-}
-
-.closed-hat-green {
-  background-color: lightskyblue;
-}
-
-.open-hat-track {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.open-hat-note {
-  border: 1px solid black;
-  min-height: 50px;
-  min-width: 50px;
-  max-height: 50px;
-  max-width: 50px;
-  cursor: pointer;
-}
-
-.open-hat-red {
-  background-color: lightcoral;
-}
-
-.open-hat-green {
-  background-color: lightblue;
-}
