--- src/SnareDrumRow/SnareDrumRow.css
@@ -1,20 +1,0 @@
-.snare-track {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-.snare-note {
-    border: 1px solid black;
-    min-height: 50px;
-    min-width: 50px;
-    cursor: pointer;
-  }
-  
-  .snare-red {
-    background-color: lightcoral;
-  }
-  
-  .snare-green {
-    background-color: lightsteelblue;
-  }
