+++ src/Square/Square.js
@@ -1,0 +1,17 @@
+const Square = ({ active, value, onClick }) => (
+  <div
+    style={{
+      display: "flex",
+      alignItems: "center",
+      justifyContent: "center",
+      width: 25,
+      height: 25,
+      background: value ? "#999" : "",
+      border: active ? "1px solid #999" : "1px solid #eee",
+      cursor: "pointer",
+    }}
+    onClick={onClick}
+  ></div>
+);
+
+export default Square;
